% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-base.R
\name{check.health}
\alias{check.health}
\title{Checking Consistency of a Health Delta-Normal Term
with a MarketRisk and a HealthRisk}
\usage{
\method{check}{health}(object, market.risk, health.risk, ...)
}
\arguments{
\item{object}{S3 object of class health.}

\item{market.risk}{S3 object of class marketRisk created using the constructor
\code{marketRisk}.}

\item{health.risk}{S3 object of class healthRisk, created using the constructor
\code{healthRisk}.}

\item{...}{additional arguments.}
}
\value{
a logical value, is the health item consistent with
  the marketRisk and the healthRisk?
}
\description{
\code{check} is a generic S3 method for classes inheriting
  from item. It is a logical method checking if the item is well defined
  with respect to a risk (i.e. that all information necessary for valuating
  the item is available).
}
\seealso{
\code{\link{check}}, \code{\link{health}}.
}
