% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-base.R
\name{check.life}
\alias{check.life}
\title{Checking Consistency of a Life Delta-Normal Remainder Term
with a MarketRisk and a HealthRisk}
\usage{
\method{check}{life}(object, market.risk, life.risk, ...)
}
\arguments{
\item{object}{S3 object of class life.}

\item{market.risk}{S3 object of class marketRisk created using the constructor
\code{marketRisk}.}

\item{life.risk}{S3 object of class lifeRisk, created using the constructor
\code{lifeRisk}.}

\item{...}{additional arguments.}
}
\value{
a logical value, is the life item consistent with
  the marketRisk and the healthRisk?
}
\description{
\code{check} is a generic S3 method for classes inheriting
  from item. It is a logical method checking if the item is well defined
  with respect to a risk (i.e. that all information necessary for valuating
  the item is available).
}
\seealso{
\code{\link{check}}, \code{\link{life}}.
}
