% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroEconomicScenarios.R
\name{check.macroEconomicScenarios}
\alias{check.macroEconomicScenarios}
\title{Checking Macro Economic Scenarios}
\usage{
\method{check}{macroEconomicScenarios}(object, market.risk, portfolio, ...)
}
\arguments{
\item{object}{an S3 object of class macroEconomicScenario.}

\item{market.risk}{an S3 object of class marketRisk.}

\item{portfolio}{an S3 object of class portfolio.}

\item{...}{additional arguments.}
}
\value{
a logical value.
}
\description{
Checking Macro Economic Scenarios
}
