% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{check}
\alias{check}
\title{Object Checks}
\usage{
check(object, ...)
}
\arguments{
\item{object}{an S3 object to check.}

\item{...}{additional parameters.}
}
\value{
a logical value.
}
\description{
\code{check} is a generic S3 method for S3 classes inheriting
  from item. It is a logical method checking if the item is well defined
  with respect to a risk (i.e. that all information necessary for valuating
  the item is available).
}
