% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-base.R
\name{check.standalone}
\alias{check.standalone}
\title{Checking Consistency of a Standalone with a MarketRisk}
\usage{
\method{check}{standalone}(object, market.risk, ...)
}
\arguments{
\item{object}{S3 object of class standalone.}

\item{market.risk}{S3 object of class marketRisk created using the constructor
\code{marketRisk}.}

\item{...}{additional arguments.}
}
\value{
a logical value, is the standalone consistent with the marketRisk?
}
\description{
\code{check} is a generic S3 method for classes inheriting from item.
  It is a logical method checking if the item is well defined with respect to
  a risk (i.e. that all information necessary for valuating the item is
  available).
}
\seealso{
\code{\link{check}}, \code{\link{standalone}}.
}
