% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthRisk-simulation.R
\name{compute.healthRisk}
\alias{compute.healthRisk}
\title{Compute a HealthRisk}
\usage{
\method{compute}{healthRisk}(object, market.risk, health.item, nsim,
  seed = NULL, ...)
}
\arguments{
\item{object}{S3 object of class healthRisk.}

\item{market.risk}{S3 object of class marketRisk created using
\code{marketRisk}.}

\item{health.item}{S3 object of class health from a portfolio.}

\item{nsim}{strictly positive integer value og length one. The number
of simulations.}

\item{seed}{positive integer value of length one. The seed for
reproducibility.}

\item{...}{additional arguments.}
}
\value{
a \code{data.table} value containing one column named \code{"healthRisk"}. The
  simulations result for a healthRisk.
}
\description{
\code{compute} is a generic S3 method for classes inheriting from risk.
  It returns a vector of aggregated simulations for the corresponding risk.
}
\seealso{
\code{\link{compute}}, \code{\link{healthRisk}}.
}
