% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroEconomicScenarios.R
\name{compute.macroEconomicScenarios}
\alias{compute.macroEconomicScenarios}
\title{Computing Macro Economic Scenarios}
\usage{
\method{compute}{macroEconomicScenarios}(object, market.risk, portfolio, ...)
}
\arguments{
\item{object}{an S3 object of class economicScenarios.}

\item{market.risk}{an S3 object of class marketRisk.}

\item{portfolio}{an S3 object of class portfolio.}

\item{...}{additional arguments.}
}
\value{
a data.table with the macro economic scenario values.
}
\description{
Computing Macro Economic Scenarios
}
