% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonLifeRisk-simulation.R
\name{compute.nonLifeRisk}
\alias{compute.nonLifeRisk}
\title{Compute a nonLifeRisk}
\usage{
\method{compute}{nonLifeRisk}(object, nsim, seed = NULL, market.risk, ...)
}
\arguments{
\item{object}{an S3 object of class nonLifeRisk.}

\item{nsim}{a strictly positive integer value og length one. The number
of simulations.}

\item{seed}{a strictly positive integer value of length one. The seed for
reproducibility.}

\item{market.risk}{an S3 object of class marketRisk created using
\code{marketRisk}.}

\item{...}{additional parameters.}
}
\value{
a data.table value containing one column named nonLifeRisk. The
  simulations result for a nonLifeRisk.
}
\description{
\code{compute} is a generic S3 method for S3 classes inheriting from risk.
  It returns a vector of aggregated simulations for the corresponding risk.
}
\seealso{
\code{\link{compute}}, \code{\link{nonLifeRisk}}.
}
