% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participationRisk-simulation.R
\name{compute.participationRisk}
\alias{compute.participationRisk}
\title{Compute a participationRisk}
\usage{
\method{compute}{participationRisk}(object, market.risk, participation.item,
  nsim, seed = NULL, ...)
}
\arguments{
\item{object}{S3 object of class participationRisk.}

\item{market.risk}{S3 object of class marketRisk.}

\item{participation.item}{S3 object of class participation.}

\item{nsim}{strictly positive integer value og length one. The number
of simulations.}

\item{seed}{positive integer value of length one. The seed for
reproducibility.}

\item{...}{additional arguments.}
}
\value{
a data.table value containing one column named participation. The
  simulations result for a participationRisk.
}
\description{
\code{compute} is a generic S3 method for S3 classes inheriting from risk.
It returns a vector of aggregated simulations for the corresponding risk.
}
\seealso{
\code{\link{compute}}, \code{\link{participationRisk}},
  \code{\link{participation}}.
}
