% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{compute}
\alias{compute}
\title{Object Computations}
\usage{
compute(object, ...)
}
\arguments{
\item{object}{an S3 object to compute.}

\item{...}{additional parameters.}
}
\value{
results of the computation.
}
\description{
\code{compute} is a generic S3 method for S3 classes
  inheriting from risk. It returns a vector of aggregated simulations
  for the corresponding risk.
}
