% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioRisk-simulation.R
\name{compute.scenarioRisk}
\alias{compute.scenarioRisk}
\title{Compute a ScenarioRisk}
\usage{
\method{compute}{scenarioRisk}(object, nsim, seed = NULL, market.risk, ...)
}
\arguments{
\item{object}{S3 object of class scenarioRisk.}

\item{nsim}{strictly positive integer value of length one. The number
of simulations.}

\item{seed}{positive integer value of length one. The seed for
reproducibility.}

\item{market.risk}{S3 object of class marketRisk created using the constructor
\code{marketRisk}.}

\item{...}{additional arguments.}
}
\value{
a \code{data.table} value containing one column named \code{"scenarioRisk"}. The
  simulations result for a scenarioRisk.
}
\description{
\code{compute} is a generic S3 method for classes inheriting from risk.
  It returns a vector of aggregated simulations for the corresponding risk.
}
\seealso{
\code{\link{compute}}, \code{\link{scenarioRisk}}.
}
