% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-valuation.R
\name{computeConstant}
\alias{computeConstant}
\title{Compute The Normalizing Constant for a log-Normal Random Variable}
\usage{
computeConstant(id, scale, cov.matrix)
}
\arguments{
\item{id}{an integer value. The risk-factor ids involved in the valuation
formula.}

\item{scale}{a numeric value. The scales corresponding to those
risk-factors.}

\item{cov.matrix}{a numeric matrix. The covariance matrix of the
risk-factors.}
}
\value{
A numeric value, the scaling constant. This is equal to \code{-0.5}
  times the variance of the linear combination of the risk-factors provided
  in the parameters.
}
\description{
This private function allows to compute scaling constants
  in the valuation formulas.
}
