% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{containsInsurance.sstOutput}
\alias{containsInsurance.sstOutput}
\title{containsInsurance Helper}
\usage{
\method{containsInsurance}{sstOutput}(object, ...)
}
\arguments{
\item{object}{sstOutput object.}

\item{...}{additional arguments.}
}
\value{
a logical value.
}
\description{
S3 generic method to check if the object contains a insuranceRisk.
}
\seealso{
\code{\link{containsInsurance}}.
}
