% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{containsLife}
\alias{containsLife}
\title{Checks if the object contains a lifeRisk.}
\usage{
containsLife(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters.}
}
\value{
a logical value.
}
\description{
S3 generic method to check if the object contains a lifeRisk.
}
\seealso{
\code{\link{containsLife}}.
}
