% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{containsLife.sstOutput}
\alias{containsLife.sstOutput}
\title{containsLife Helper}
\usage{
\method{containsLife}{sstOutput}(object, ...)
}
\arguments{
\item{object}{sstOutput object.}

\item{...}{additional arguments.}
}
\value{
a logical value.
}
\description{
S3 generic method to check if the object contains a lifeRisk.
}
\seealso{
\code{\link{containsLife}}.
}
