% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{containsMarket.sstOutput}
\alias{containsMarket.sstOutput}
\title{containsMarket Helper}
\usage{
\method{containsMarket}{sstOutput}(object, ...)
}
\arguments{
\item{object}{sstOutput object.}

\item{...}{additional arguments.}
}
\value{
a logical value.
}
\description{
S3 generic method to check if the object contains a MarketRisk.
}
\seealso{
\code{\link{containsMarket}}.
}
