% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{containsNonLife.sstOutput}
\alias{containsNonLife.sstOutput}
\title{containsNonLife Helper}
\usage{
\method{containsNonLife}{sstOutput}(object, ...)
}
\arguments{
\item{object}{sstOutput object.}

\item{...}{additional arguments.}
}
\value{
a logical value.
}
\description{
S3 generic method to check if the object
  contains nonLifeRisk.
}
\seealso{
\code{\link{containsNonLife}}.
}
