% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{containsParticipation.sstOutput}
\alias{containsParticipation.sstOutput}
\title{containsParticipation Helper}
\usage{
\method{containsParticipation}{sstOutput}(object, ...)
}
\arguments{
\item{object}{sstOutput object.}

\item{...}{additional arguments.}
}
\value{
a logical value.
}
\description{
S3 generic method to check if the object
  contains participation.
}
\seealso{
\code{\link{containsParticipation}}.
}
