% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{containsScenario.sstOutput}
\alias{containsScenario.sstOutput}
\title{containsScenario Helper}
\usage{
\method{containsScenario}{sstOutput}(object, ...)
}
\arguments{
\item{object}{sstOutput object.}

\item{...}{additional arguments.}
}
\value{
a logical value.
}
\description{
S3 generic method to check if the object
  contains scenario.
}
\seealso{
\code{\link{containsScenario}}.
}
