% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{creditRisk}
\alias{creditRisk}
\title{Credit risk}
\usage{
creditRisk(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get credit risk.
}
