% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-figures.R
\name{creditRisk.sstOutput}
\alias{creditRisk.sstOutput}
\title{Get Credit Risk from sstOutput}
\usage{
\method{creditRisk}{sstOutput}(object, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{...}{additional parameters.}
}
\value{
a numeric value. The credit risk.
}
\description{
S3 method to extract the credit risk from an sstOutput.
}
\seealso{
\code{\link{creditRisk}}.
}
