% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{currencyIsIn}
\alias{currencyIsIn}
\title{Currency in Object?}
\usage{
currencyIsIn(object, ...)
}
\arguments{
\item{object}{an S3 object potentially containing the currency.}

\item{...}{additional parameters.}
}
\value{
a logical value.
}
\description{
S3 generic to check that the currency is in
  the object.
}
