% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-base.R
\name{currencyIsIn.standalone}
\alias{currencyIsIn.standalone}
\title{Currency in Standalone?}
\usage{
\method{currencyIsIn}{standalone}(object, from, to, ...)
}
\arguments{
\item{object}{S3 object of class standalone.}

\item{from}{a character value of length one. A currency.}

\item{to}{a character value of length one. A currency.}

\item{...}{additional arguments.}
}
\value{
a logical value, is the currency in the standalone?
}
\description{
S3 generic to check that the currency is in
  the object.
}
\seealso{
\code{\link{currencyIsIn}}.
}
