% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{equityIsIn}
\alias{equityIsIn}
\title{Equity in Object?}
\usage{
equityIsIn(object, ...)
}
\arguments{
\item{object}{an S3 object potentially containing the equity.}

\item{...}{additional parameters.}
}
\value{
a logical value.
}
\description{
S3 generic to check that the equity is in
  the object.
}
