% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-base.R
\name{equityIsIn.standalone}
\alias{equityIsIn.standalone}
\title{Equity in Standalone?}
\usage{
\method{equityIsIn}{standalone}(object, type, currency, ...)
}
\arguments{
\item{object}{S3 object of class standalone.}

\item{type}{character value of length one. The type of an asset.}

\item{currency}{character value of length one. The currency of the asset.}

\item{...}{additional arguments.}
}
\value{
a logical value, is the equity in the standalone?
}
\description{
S3 generic to check that the equity is in
  the object.
}
\seealso{
\code{\link{equityIsIn}}, \code{\link{asset}}.
}
