% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelToSstModel.R
\name{excelToSstModel}
\alias{excelToSstModel}
\title{Parsing an Excel Template to sstModel}
\usage{
excelToSstModel(path, with.log = F)
}
\arguments{
\item{path}{a character value. A valid path of an input excel workbook.
The path can be relative or not.}

\item{with.log}{logical value. Should the error/warning-log be returned?}
}
\value{
an S3 object of class sstModel, built from the input
  fundamenental data sheets.
}
\description{
this function is intended to parse the excel template
  provided by FINMA into an sstModel.
}
\seealso{
\code{\link{sstModel}}.
}
