% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-base.R
\name{format.asset}
\alias{format.asset}
\title{Formating an Asset with Direct Market Price}
\usage{
\method{format}{asset}(x, ...)
}
\arguments{
\item{x}{S3 object of class asset.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
format method for the S3 class asset.
}
\seealso{
\code{\link[base]{format}}, \code{\link{asset}}.
}
