% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetForward-base.R
\name{format.assetForward}
\alias{format.assetForward}
\title{Formating an Index-Forward}
\usage{
\method{format}{assetForward}(x, ...)
}
\arguments{
\item{x}{S3 object of class assetForward.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
format method for the S3 class assetForward.
}
\seealso{
\code{\link[base]{format}}, \code{\link{assetForward}}.
}
