% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow-base.R
\name{format.cashflow}
\alias{format.cashflow}
\title{Formating a Fixed-Income-Asset}
\usage{
\method{format}{cashflow}(x, ...)
}
\arguments{
\item{x}{S3 object of class cashflow.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a Fixed-Income-Asset
}
\seealso{
\code{\link[base]{format}}, \code{\link{cashflow}}
}
