% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-base.R
\name{format.delta}
\alias{format.delta}
\title{Formating a Delta-Normal Remainder Term}
\usage{
\method{format}{delta}(x, ...)
}
\arguments{
\item{x}{S3 object of class delta.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a Delta-Normal Remainder Term
}
\seealso{
\code{\link[base]{format}}, \code{\link{delta}}.
}
