% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxForward-base.R
\name{format.fxForward}
\alias{format.fxForward}
\title{Formating an FX-Forward}
\usage{
\method{format}{fxForward}(x, ...)
}
\arguments{
\item{x}{an S3 object of class fxForward.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
format method for the S3 class fxForward.
}
\seealso{
\code{\link[base]{format}}, \code{\link{fxForward}}.
}
