% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-base.R
\name{format.health}
\alias{format.health}
\title{Formating a Health Delta-Normal Term}
\usage{
\method{format}{health}(x, ...)
}
\arguments{
\item{x}{S3 object of class health.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a Health Delta-Normal Term
}
\seealso{
\code{\link[base]{format}}, \code{\link{health}}.
}
