% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthRisk-base.R
\name{format.healthRisk}
\alias{format.healthRisk}
\title{Formating a HealhRisk}
\usage{
\method{format}{healthRisk}(x, ...)
}
\arguments{
\item{x}{S3 object of class healthRisk.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a HealhRisk
}
\seealso{
\code{\link[base]{format}}, \code{\link{healthRisk}}.
}
