% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liability-base.R
\name{format.liability}
\alias{format.liability}
\title{Formating an Insurance Liability}
\usage{
\method{format}{liability}(x, ...)
}
\arguments{
\item{x}{S3 object of class liability.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating an Insurance Liability
}
\seealso{
\code{\link[base]{format}}, \code{\link{liability}}.
}
