% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-base.R
\name{format.life}
\alias{format.life}
\title{Formating a Life Delta-Normal Remainder Term}
\usage{
\method{format}{life}(x, ...)
}
\arguments{
\item{x}{S3 object of class life.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a Life Delta-Normal Remainder Term
}
\seealso{
\code{\link[base]{format}}, \code{\link{life}}.
}
