% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeRisk-base.R
\name{format.lifeRisk}
\alias{format.lifeRisk}
\title{Formating a LifeRisk}
\usage{
\method{format}{lifeRisk}(x, ...)
}
\arguments{
\item{x}{an S3 object of class lifeRisk.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a LifeRisk
}
\seealso{
\code{\link[base]{format}}, \code{\link{lifeRisk}}.
}
