% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{format.marketRisk}
\alias{format.marketRisk}
\title{Formating a marketRisk}
\usage{
\method{format}{marketRisk}(x, ...)
}
\arguments{
\item{x}{S3 object of class marketRisk.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
format method for S3 class marketRisk.
}
\seealso{
\code{\link[base]{format}}, \code{\link{marketRisk}}.
}
