% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonLifeRisk-base.R
\name{format.nonLifeRisk}
\alias{format.nonLifeRisk}
\title{Formating a nonLifeRisk}
\usage{
\method{format}{nonLifeRisk}(x, ...)
}
\arguments{
\item{x}{S3 object of class nonLifeRisk.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a nonLifeRisk
}
\seealso{
\code{\link[base]{format}}, \code{\link{nonLifeRisk}}.
}
