% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participation-base.R
\name{format.participation}
\alias{format.participation}
\title{Formating a Participation}
\usage{
\method{format}{participation}(x, ...)
}
\arguments{
\item{x}{S3 object of the class participation.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a Participation
}
\seealso{
\code{\link[base]{format}}, \code{\link{participation}}.
}
