% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participationRisk-base.R
\name{format.participationRisk}
\alias{format.participationRisk}
\title{Formating a ParticipationRisk}
\usage{
\method{format}{participationRisk}(x, ...)
}
\arguments{
\item{x}{an S3 object of class participationRisk.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
format method for S3 class participationRisk.
}
\seealso{
\code{\link[base]{format}}, \code{\link{participationRisk}}.
}
