% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio-base.R
\name{format.portfolio}
\alias{format.portfolio}
\title{Formating a Portfolio}
\usage{
\method{format}{portfolio}(x, ...)
}
\arguments{
\item{x}{S3 object of class portfolio.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a Portfolio
}
\seealso{
\code{\link{format}}, \code{\link{portfolio}}.
}
