% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioRisk-base.R
\name{format.scenarioRisk}
\alias{format.scenarioRisk}
\title{Formating a ScenarioRisk}
\usage{
\method{format}{scenarioRisk}(x, ...)
}
\arguments{
\item{x}{S3 object of class scenarioRisk.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a ScenarioRisk
}
\seealso{
\code{\link{format}}, \code{\link{scenarioRisk}}.
}
