% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstModel-base.R
\name{format.sstModel}
\alias{format.sstModel}
\title{Formating a sstModel}
\usage{
\method{format}{sstModel}(x, ...)
}
\arguments{
\item{x}{S3 object of class sstModel.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a sstModel
}
\seealso{
\code{\link[base]{format}}, \code{\link{sstModel}}.
}
