% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{format.sstOutput}
\alias{format.sstOutput}
\title{Formating a sstOutput}
\usage{
\method{format}{sstOutput}(x, ...)
}
\arguments{
\item{x}{S3 object of class sstOutput.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a sstOutput
}
\seealso{
\code{\link[base]{format}}.
}
