% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-base.R
\name{format.standalone}
\alias{format.standalone}
\title{Formating a standalone}
\usage{
\method{format}{standalone}(x, ...)
}
\arguments{
\item{x}{S3 object of class standalone.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a standalone
}
\seealso{
\code{\link[base]{format}}, \code{\link{standalone}}.
}
