% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio-base.R
\name{format.summary.portfolio}
\alias{format.summary.portfolio}
\title{Formating a Summary of Portfolio}
\usage{
\method{format}{summary.portfolio}(x, ...)
}
\arguments{
\item{x}{an S3 object of class summary.portfolio.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
Formating a Summary of Portfolio
}
\seealso{
\code{\link[base]{format}}
}
