% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstModel-base.R
\name{format.summary.sstModel}
\alias{format.summary.sstModel}
\title{Formating a Summary of sstModel}
\usage{
\method{format}{summary.sstModel}(x, ...)
}
\arguments{
\item{x}{an S3 object of class summary.sstModel.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
Formating a Summary of sstModel
}
\seealso{
\code{\link[base]{format}}
}
