% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{format.summary.sstOutput}
\alias{format.summary.sstOutput}
\title{Formating a Summary of sstOutput}
\usage{
\method{format}{summary.sstOutput}(x, ...)
}
\arguments{
\item{x}{S3 object of class summary.sstOutput.}

\item{...}{additional arguments.}
}
\value{
a character value.
}
\description{
Formating a Summary of sstOutput
}
\seealso{
\code{\link[base]{format}}.
}
