% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelToSstModel.R
\name{generateError}
\alias{generateError}
\title{Generate error message from an error log}
\usage{
generateError(error.log, warning.log, line.break = "\\n ")
}
\arguments{
\item{error.log}{a data.frame with following fields:
\itemize{
  \item \code{sheet}: character, the sheet name.
  \item \code{row}: integer, the row position.
  \item \code{column}: integer, the column position.
  \item \code{message}: character, the error message.
}}

\item{warning.log}{a data.frame similar to \code{error.log}.}

\item{line.break}{a character value, separation between error messages.}
}
\value{
a character value, the corresponding error message.
}
\description{
this function transforms an error log into an error message.
}
\seealso{
\code{\link{excelToSstModel}}.
}
