% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{generateExpression}
\alias{generateExpression}
\title{Generate an Expression}
\usage{
generateExpression(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional arguments.}
}
\value{
an expression.
}
\description{
method to generate an expression.
}
