% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{generateFunction}
\alias{generateFunction}
\title{Generate a Function}
\usage{
generateFunction(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional arguments.}
}
\value{
a function.
}
\description{
method to generate a function.
}
