% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getCurrencyName}
\alias{getCurrencyName}
\title{Get A Currency Name}
\usage{
getCurrencyName(object, ...)
}
\arguments{
\item{object}{an S3 object containg the currency.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
S3 generic to get a currency name.
}
