% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getCurrencyScale.marketRisk}
\alias{getCurrencyScale.marketRisk}
\title{Get A Currency Scale}
\usage{
\method{getCurrencyScale}{marketRisk}(object, from, to, ...)
}
\arguments{
\item{object}{S3 object of class marketRisk.}

\item{from}{character value. A well defined type in
\code{object}.}

\item{to}{character value. A well defined currency in
\code{object} for the departure currency \code{from}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a currency scale.
}
\note{
This method is private and does not test validity or coherence of its
  arguments.
}
\seealso{
\code{\link{getCurrencyScale}}.
}
