% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getCurrencyScale}
\alias{getCurrencyScale}
\title{Get A Currency Scale}
\usage{
getCurrencyScale(object, ...)
}
\arguments{
\item{object}{an S3 object containg the currency.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a currency scale.
}
