% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getDeltaId.marketRisk}
\alias{getDeltaId.marketRisk}
\title{Get A Delta ID}
\usage{
\method{getDeltaId}{marketRisk}(object, name, ...)
}
\arguments{
\item{object}{S3 object of class marketRisk.}

\item{name}{character value. A well defined risk factor names
in \code{object}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a delta id.
}
\note{
This method is private and does not test validity
  or coherence of its arguments.
}
\seealso{
\code{\link{getDeltaId}}.
}
