% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getDeltaId}
\alias{getDeltaId}
\title{Get A Delta ID}
\usage{
getDeltaId(object, ...)
}
\arguments{
\item{object}{an S3 object containg the delta.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a delta id.
}
