% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getDrbc}
\alias{getDrbc}
\title{Get drbc}
\usage{
getDrbc(object, with.scenario = F, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{with.scenario}{a logical value.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get drbc
}
