% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{getDrbc.sstOutput}
\alias{getDrbc.sstOutput}
\title{Get drbc}
\usage{
\method{getDrbc}{sstOutput}(object, with.scenario = F, exp.shortfall = F,
  ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{with.scenario}{logical value.}

\item{exp.shortfall}{logical value, by default set to \code{FALSE}.
Should the expected shortfall be returned?}

\item{...}{additional arguments.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get drbc
}
\seealso{
\code{\link{getDrbc}}.
}
