% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getEquityId.marketRisk}
\alias{getEquityId.marketRisk}
\title{Get An Equity ID}
\usage{
\method{getEquityId}{marketRisk}(object, type, currency, ...)
}
\arguments{
\item{object}{S3 object of class marketRisk.}

\item{type}{character value. A well defined type in
\code{object}.}

\item{currency}{character value. A well defined
currency in \code{object} for the type \code{type}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get an equity id.
}
\note{
This method is private and does not test validity or coherence of its
  arguments.
}
\seealso{
\code{\link{getEquityId}}.
}
