% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getEquityName}
\alias{getEquityName}
\title{Get An Equity Name}
\usage{
getEquityName(object, ...)
}
\arguments{
\item{object}{an S3 object containg the equity.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
S3 generic to get an equity name.
}
