% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getEquityScale}
\alias{getEquityScale}
\title{Get An Equity Scale}
\usage{
getEquityScale(object, ...)
}
\arguments{
\item{object}{an S3 object containg the equity.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get an equity Scale.
}
