% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getHealthQuantile}
\alias{getHealthQuantile}
\title{Get A Health Item Quantile}
\usage{
getHealthQuantile(object, ...)
}
\arguments{
\item{object}{an S3 object containg the health item.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a health item quantile.
}
