% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{getHealthRisk.sstOutput}
\alias{getHealthRisk.sstOutput}
\title{Get Health Insurance Risk}
\usage{
\method{getHealthRisk}{sstOutput}(object, exp.shortfall = F, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{exp.shortfall}{logical value, by default set to \code{FALSE}.
Should the expected shortfall be returned?}

\item{...}{additional arguments.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get health insurance risk.
}
\seealso{
\code{\link{getInsuranceRisk}}.
}
