% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getInitialFX.marketRisk}
\alias{getInitialFX.marketRisk}
\title{Get An Initial FX}
\usage{
\method{getInitialFX}{marketRisk}(object, from, to, ...)
}
\arguments{
\item{object}{S3 object of class marketRisk.}

\item{from}{character value. A well-defined currency defined in
\code{object}.}

\item{to}{character value. A well-defined currency defined in
\code{object}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get initial fx.
}
\note{
This method is private and does not test validity or coherence of its
  arguments.
}
\seealso{
\code{\link{getInitialFX}}.
}
