% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getInitialFX}
\alias{getInitialFX}
\title{Get An Initial FX}
\usage{
getInitialFX(object, ...)
}
\arguments{
\item{object}{an S3 object containg initial fx.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get initial fx.
}
