% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getInitialRate.marketRisk}
\alias{getInitialRate.marketRisk}
\title{Get An Initial Rate}
\usage{
\method{getInitialRate}{marketRisk}(object, time, currency, ...)
}
\arguments{
\item{object}{object of class marketRisk.}

\item{time}{integer value. A well defined time in object for
\code{currency}.}

\item{currency}{character value. A well defined currency in
\code{object}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get initial rate.
}
\note{
This method is private and does not test validity or coherence of its
  arguments.
}
\seealso{
\code{\link{getInitialRate}}.
}
