% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getInitialRate}
\alias{getInitialRate}
\title{Get An Initial Rate}
\usage{
getInitialRate(object, ...)
}
\arguments{
\item{object}{an S3 object containg initial rate.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get initial rate.
}
