% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getInitialSpread}
\alias{getInitialSpread}
\title{Get An Initial Spread}
\usage{
getInitialSpread(object, ...)
}
\arguments{
\item{object}{an S3 object containg initial spread.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get initial spread.
}
