% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeRisk-base.R
\name{getLifeQuantile.lifeRisk}
\alias{getLifeQuantile.lifeRisk}
\title{Get LifeRisk Quantiles}
\usage{
\method{getLifeQuantile}{lifeRisk}(object, name, ...)
}
\arguments{
\item{object}{an S3 object of class lifeRisk.}

\item{name}{a character value. A well defined risk factor
names in object.}

\item{...}{additional arguments.}
}
\value{
a numeric value.
}
\description{
This method is private and does not test
  validity or coherence of its arguments.
}
\seealso{
\code{\link{getLifeQuantile}}, \code{\link{lifeRisk}}.
}
