% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getLifeQuantile}
\alias{getLifeQuantile}
\title{Get A Life Item Quantile}
\usage{
getLifeQuantile(object, ...)
}
\arguments{
\item{object}{an S3 object containg the life item.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a life item quantile.
}
