% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getLifeRisk}
\alias{getLifeRisk}
\title{Get Life Risk}
\usage{
getLifeRisk(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get life insurance risk.
}
