% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getMappingTime.marketRisk}
\alias{getMappingTime.marketRisk}
\title{Get A Time Mapping}
\usage{
\method{getMappingTime}{marketRisk}(object, time, ...)
}
\arguments{
\item{object}{S3 object of class marketRisk.}

\item{time}{integer value. A well defined time
in \code{object}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a time mapping.
}
\note{
This method is private and does not test validity or coherence of its
  arguments.
}
\seealso{
\code{\link{getMappingTime}}.
}
