% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getMappingTime}
\alias{getMappingTime}
\title{Get A Time Mapping}
\usage{
getMappingTime(object, ...)
}
\arguments{
\item{object}{an S3 object containg a time mapping.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
S3 generic to get a time mapping.
}
