% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getMarketParticipationRisk}
\alias{getMarketParticipationRisk}
\title{Get Aggregated Market Risk and Participation}
\usage{
getMarketParticipationRisk(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get aggregated market risk
  and participation.
}
