% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{getMarketParticipationRisk.sstOutput}
\alias{getMarketParticipationRisk.sstOutput}
\title{Get Aggregated Market and Participation Risk}
\usage{
\method{getMarketParticipationRisk}{sstOutput}(object, exp.shortfall = F, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{exp.shortfall}{logical value, by default set to \code{FALSE}.
Should the expected shortfall be returned?}

\item{...}{additional arguments.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get aggregated market risk
  and participation.
}
\seealso{
\code{\link{getMarketRisk}}.
}
