% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getMarketRisk}
\alias{getMarketRisk}
\title{Get Market Risk}
\usage{
getMarketRisk(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get market risk.
}
