% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getParticipation}
\alias{getParticipation}
\title{Get Participation}
\usage{
getParticipation(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get participation.
}
