% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getRateId.marketRisk}
\alias{getRateId.marketRisk}
\title{Get A Rate ID}
\usage{
\method{getRateId}{marketRisk}(object, currency, horizon, ...)
}
\arguments{
\item{object}{S3 object of class marketRisk.}

\item{currency}{character value. A well defined currency in
\code{object}.}

\item{horizon}{character value. A well defined horizon in
\code{object} for the departure currency \code{currency}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a rate id.
}
\note{
This method is private and does not test validity or coherence of its
  arguments.
}
\seealso{
\code{\link{getRateId}}.
}
