% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getRateId}
\alias{getRateId}
\title{Get A Rate ID}
\usage{
getRateId(object, ...)
}
\arguments{
\item{object}{an S3 object containg the rate.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a rate id.
}
