% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getRateName}
\alias{getRateName}
\title{Get A Rate Name}
\usage{
getRateName(object, ...)
}
\arguments{
\item{object}{an S3 object containg the rate.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
S3 generic to get a rate name.
}
