% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getRateScale}
\alias{getRateScale}
\title{Get A Rate Scale}
\usage{
getRateScale(object, ...)
}
\arguments{
\item{object}{an S3 object containg the rate.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a rate scale.
}
