% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getScenarioRisk}
\alias{getScenarioRisk}
\title{Get Scenario Risk}
\usage{
getScenarioRisk(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get scenario risk.
}
