% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{getScenarioRisk.sstOutput}
\alias{getScenarioRisk.sstOutput}
\title{Get Scenario Risk}
\usage{
\method{getScenarioRisk}{sstOutput}(object, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{...}{additional arguments.}
}
\value{
a numeric value.
}
\description{
S3 generic method to get scenario risk.
}
\seealso{
\code{\link{getScenarioRisk}}.
}
