% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getSpreadId.marketRisk}
\alias{getSpreadId.marketRisk}
\title{Get A Spread ID}
\usage{
\method{getSpreadId}{marketRisk}(object, currency, rating, ...)
}
\arguments{
\item{object}{an S3 object of class marketRisk.}

\item{currency}{a character value. A well defined currency in
\code{object}.}

\item{rating}{a character value. A well defined rating in
\code{object} for the currency \code{currency}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a spread id.
}
\note{
This method is private and does not test validity or coherence of its
  arguments.
}
\seealso{
\code{\link{getSpreadId}}.
}
