% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getSpreadName.marketRisk}
\alias{getSpreadName.marketRisk}
\title{Get A Spread Name}
\usage{
\method{getSpreadName}{marketRisk}(object, currency, rating, ...)
}
\arguments{
\item{object}{an S3 object of class marketRisk.}

\item{currency}{a character value. A well defined currency in
\code{object}.}

\item{rating}{a character value. A well defined rating in
\code{object} for the currency \code{currency}.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
S3 generic to get a spread name.
}
\note{
This method is private and does not test validity or coherence of its
  arguments.
}
\seealso{
\code{\link{getSpreadName}}.
}
