% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getSpreadName}
\alias{getSpreadName}
\title{Get A Spread Name}
\usage{
getSpreadName(object, ...)
}
\arguments{
\item{object}{an S3 object containg the spread.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
S3 generic to get a spread name.
}
