% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{getSpreadScale.marketRisk}
\alias{getSpreadScale.marketRisk}
\title{Get A Spread Scale}
\usage{
\method{getSpreadScale}{marketRisk}(object, currency, rating, ...)
}
\arguments{
\item{object}{S3 object of class marketRisk.}

\item{currency}{character value. A well defined currency in
\code{object}.}

\item{rating}{character value. A well defined rating in
\code{object} for the currency \code{currency}.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a spread scale.
}
\note{
This method is private and does not test validity
   or coherence of its arguments.
}
\seealso{
\code{\link{getSpreadScale}}.
}
