% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{getSpreadScale}
\alias{getSpreadScale}
\title{Get A Spread Scale}
\usage{
getSpreadScale(object, ...)
}
\arguments{
\item{object}{an S3 object containg the spread.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic to get a spread scale.
}
