% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{initialFX}
\alias{initialFX}
\title{Constructing initial FX Rates}
\usage{
initialFX(from, to, fx)
}
\arguments{
\item{from}{character value, the currencies.}

\item{to}{character value, the currencies.}

\item{fx}{numeric value, the fx rates.}
}
\value{
a data.frame with option \code{stringsAsFactors = FALSE}.
}
\description{
Constructor for initial FX values.
}
\seealso{
\code{\link{marketRisk}}.
}
