% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{initialRate}
\alias{initialRate}
\title{Constructing Initial Interest Rates}
\usage{
initialRate(time, currency, rate)
}
\arguments{
\item{time}{integer value, the times to maturity.}

\item{currency}{character value, the currencies.}

\item{rate}{numeric value, the interest rates.}
}
\value{
a data.frame with option \code{stringsAsFactors = FALSE}.
}
\description{
Constructor for initial Initial Rates values.
}
\seealso{
\code{\link{marketRisk}}.
}
