% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-general.R
\name{initialSpread}
\alias{initialSpread}
\title{Compute Initial Spread}
\usage{
initialSpread(market.value, times, coupons, risk.free, ...)
}
\arguments{
\item{market.value}{a numeric value, the total market value for the bond.}

\item{times}{a numeric vector, the times of the coupons.}

\item{coupons}{a numeric vector, the corresponding coupon cash flows.}

\item{risk.free}{a numeric vector, the corresponding risk-free rates with
continuous compounding.}

\item{...}{additional parameters to be passed to \code{newtonRaphson}.}
}
\value{
a numeric value, the corresponding spread.
}
\description{
compute initial spread of a bond from its market value.
}
