% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-general.R
\name{intToGroups}
\alias{intToGroups}
\title{Ordered Vector of Integers to List of consecutive integers}
\usage{
intToGroups(x)
}
\arguments{
\item{x}{a vector of integers.}
}
\value{
a list of integer vectors.
}
\description{
this helper function helps to group consecutive integers
in a sequence of integers.
}
