% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.assetForward}
\alias{is.assetForward}
\title{Assess Class Membership (assetForward S3 class)}
\usage{
is.assetForward(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{assetForward}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{assetForward}}
}
