% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.cashflow}
\alias{is.cashflow}
\title{Assess Class Membership (cashflow S3 class)}
\usage{
is.cashflow(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{cashflow}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{cashflow}}.
}
