% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.delta}
\alias{is.delta}
\title{Assess Class Membership (delta S3 class)}
\usage{
is.delta(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{delta}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{delta}}.
}
