% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.equity}
\alias{is.equity}
\title{Assess Class Membership (equity S3 class)}
\usage{
is.equity(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{equity}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{is.riskFactor}}.
}
