% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.fxForward}
\alias{is.fxForward}
\title{Assess Class Membership (fxForward S3 class)}
\usage{
is.fxForward(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{fxForward}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{fxForward}}.
}
