% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.health}
\alias{is.health}
\title{Assess Class Membership (health S3 class)}
\usage{
is.health(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{health}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{health}}.
}
