% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.healthRisk}
\alias{is.healthRisk}
\title{Assess Class Membership (healthRisk S3 class)}
\usage{
is.healthRisk(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{healthRisk}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{healthRisk}}.
}
