% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.insuranceItem}
\alias{is.insuranceItem}
\title{Assess Class Membership (insuranceItem S3 class)}
\usage{
is.insuranceItem(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{insuranceItem}.
}
\description{
Functions to test inheritance relationships.
}
