% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.insuranceRisk}
\alias{is.insuranceRisk}
\title{Assess Class Membership (insuranceRisk S3 class)}
\usage{
is.insuranceRisk(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{insuranceRisk}.
}
\description{
Functions to test inheritance relationships.
}
