% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.item}
\alias{is.item}
\title{Assess Class Membership (item S3 class)}
\usage{
is.item(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{item}.
}
\description{
Function to test inheritance relationships.
}
