% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.liability}
\alias{is.liability}
\title{Assess Class Membership (liability S3 class)}
\usage{
is.liability(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{liability}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{liability}}
}
