% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.life}
\alias{is.life}
\title{Assess Class Membership (life S3 class)}
\usage{
is.life(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{life}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{life}}.
}
