% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.lifeRisk}
\alias{is.lifeRisk}
\title{Assess Class Membership (lifeRisk S3 class)}
\usage{
is.lifeRisk(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{lifeRisk}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{lifeRisk}}.
}
