% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.macroEconomicScenarios}
\alias{is.macroEconomicScenarios}
\title{Assess Class Membership (macroEconomicScenarios S3 class)}
\usage{
is.macroEconomicScenarios(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{macroEconomicScenarios}.
}
\description{
Functions to test inheritance relationships.
}
