% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.marketItem}
\alias{is.marketItem}
\title{Assess Class Membership (marketItem S3 class)}
\usage{
is.marketItem(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{marketItem}.
}
\description{
Function to test inheritance relationships.
}
