% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.marketRisk}
\alias{is.marketRisk}
\title{Assess Class Membership (marketRisk S3 class)}
\usage{
is.marketRisk(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{marketRisk}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{marketRisk}}.
}
