% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.nonLifeRisk}
\alias{is.nonLifeRisk}
\title{Assess Class Membership (nonLifeRisk S3 class)}
\usage{
is.nonLifeRisk(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{nonLifeRisk}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{nonLifeRisk}}.
}
