% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.participation}
\alias{is.participation}
\title{Assess Class Membership (participation S3 class)}
\usage{
is.participation(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{participation}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{participation}}.
}
