% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.participationRisk}
\alias{is.participationRisk}
\title{Assess Class Membership (standalone S3 class)}
\usage{
is.participationRisk(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{participationRisk}.
}
\description{
Functions to test inheritance relationships.
}
