% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.portfolio}
\alias{is.portfolio}
\title{Assess Class Membership (portfolio S3 class)}
\usage{
is.portfolio(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{portfolio}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{portfolio}}.
}
