% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.rate}
\alias{is.rate}
\title{Assess Class Membership (rate S3 class)}
\usage{
is.rate(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{rate}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{is.riskFactor}}.
}
