% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.risk}
\alias{is.risk}
\title{Assess Class Membership (risk S3 class)}
\usage{
is.risk(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{risk}.
}
\description{
Functions to test inheritance relationships.
}
