% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.riskFactor}
\alias{is.riskFactor}
\title{Assess Class Membership (riskFactor S3 class)}
\usage{
is.riskFactor(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{riskFactor}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{currency}}, \code{\link{rate}},
\code{\link{spread}}, \code{\link{equity}}, \code{\link{pcRate}}.
}
