% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.scenarioRisk}
\alias{is.scenarioRisk}
\title{Assess Class Membership (scenerioRisk S3 class)}
\usage{
is.scenarioRisk(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{scenerioRisk}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{scenarioRisk}}.
}
