% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.sstModel}
\alias{is.sstModel}
\title{Assess Class Membership (sstModel S3 class)}
\usage{
is.sstModel(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{sstModel}.
}
\description{
Functions to test inheritance relationships.
}
\seealso{
\code{\link{sstModel}}.
}
