% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-class.R
\name{is.sstOutput}
\alias{is.sstOutput}
\title{Assess Class Membership (sstOutput S3 class)}
\usage{
is.sstOutput(x)
}
\arguments{
\item{x}{an S3 object.}
}
\value{
a logical value which indicates membership of class \code{sstOutput}.
}
\description{
Functions to test inheritance relationships.
}
