% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-valuation.R
\name{itemListToExpression}
\alias{itemListToExpression}
\title{Item List to Valuation Expression Helper}
\usage{
itemListToExpression(item.list, market.item.types, market.risk,
  standalone = NULL)
}
\arguments{
\item{item.list}{a list of marketItem S3 objects.}

\item{market.item.types}{a character value representing
a subset of marketItem classes.}

\item{market.risk}{a marketRisk S3 object.}

\item{standalone}{an S3 object of class standalone.}
}
\value{
a character value representing the aggregated valuation expression.
}
\description{
helper function to convert a list
 of market.items to an aggregated valuation expression.
}
