% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-valuation.R
\name{itemListToFunction}
\alias{itemListToFunction}
\title{Item List to Valuation Function Helper}
\usage{
itemListToFunction(item.list, market.item.types, market.risk,
  with.constant = F)
}
\arguments{
\item{item.list}{a list of marketItem S3 objects.}

\item{market.item.types}{a character value representing
a subset of marketItem classes.}

\item{market.risk}{a marketRisk S3 object.}

\item{with.constant}{a logical value. Should the expression be with constant or not?}
}
\value{
a function representing the aggregated valuation function.
}
\description{
helper function to convert a list
 of market.items to an aggregated valuation function.
}
