% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelToSstModel.R
\name{keywordToValue}
\alias{keywordToValue}
\title{Extract a value from the excel template}
\usage{
keywordToValue(path, keyword, mapping.values)
}
\arguments{
\item{path}{a character value. A valid path of an input excel workbook.
The path can be relative or not.}

\item{keyword}{a character value. A valid keyword corresponding to a cell.}

\item{mapping.values}{a data.frame with three columns:
\itemize{
  \item \code{keyword}: character, list of keywords.
  \item \code{name}: character, sheet names corresponding to the keywords.
  \item \code{row}: integer, row number corresponding to position in
    the excel sheet.
  \item \code{col}: integer, column number corresponding to position in
    the excel sheet.
}}
}
\value{
the value of the corresponding cell.
}
\description{
this function extracts single values from the excel input
workbook.
}
\seealso{
\code{\link{excelToSstModel}}.
}
