% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchDashboard.R
\name{launchDashboard}
\alias{launchDashboard}
\title{Launching The Dashboard In A Browser}
\usage{
launchDashboard()
}
\value{
None (intended for side-effects)
}
\description{
This function launch an interative dashboard for
  SST computations.
}
