% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-valuation.R
\name{logNormalExpression}
\alias{logNormalExpression}
\title{Log-Normal Expression Helper}
\usage{
logNormalExpression(object, market.risk, standalone)
}
\arguments{
\item{object}{a S3 object of class item.'}

\item{market.risk}{a S3 object of class marketRisk.}

\item{standalone}{a S3 object of class standalone.}
}
\value{
a character value.
}
\description{
This private function creates a log-normal expression.
}
