% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroEconomicScenarios.R
\name{macroEconomicScenarios}
\alias{macroEconomicScenarios}
\title{Constructing Macro Economic Scenarios}
\usage{
macroEconomicScenarios(macro.economic.scenario.table)
}
\arguments{
\item{macro.economic.scenario.table}{a numeric matrix with named columns and rows. Each row represents
a different economic scenario and each column is associated
to a risk-factor appearing in a marketRisk (or a participation contained in the portfolio).
The rownames of the matrix should indicate the names
of the economic scenarios and the columns names should exactly match the names of the base risk-factors defined in a
a marketRisk (respecting the order). In addition, if the underlying portfolio also contains a participation,
an additional column named "participation" should be included in this table as the last column.}
}
\value{
an S3 object of class macroEconomicScenarios.
}
\description{
\code{macroEconomicScenario} is an S3 method to construct
  macro economic scenarios, i.e. constrained values taken by the market risk-factors and potentially
  participation.
}
