% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFactor.R
\name{mappingTable}
\alias{mappingTable}
\title{Constructing a Mapping Table}
\usage{
mappingTable(..., list.arg = F)
}
\arguments{
\item{...}{riskFactor objects. Please note that no risk factor name can be chosen
among the following reserved words (in that case it would trigger an error):
\itemize{
  \item \code{marketRisk}
  \item \code{lifeRisk}
  \item \code{healthRisk}
  \item \code{nonLifeRisk}
  \item \code{scenarioRisk}
  \item \code{participationRisk}
  \item \code{participation}
  \item \code{marketParticipationRisk}
  \item \code{asset}
  \item \code{cashflow}
  \item \code{liability}
  \item \code{assetForward}
  \item \code{fxForward}
  \item \code{delta}
}}

\item{list.arg}{a logical value, by default set to \code{FALSE}.
It allows to use \code{...} argument to pass a list of objects of class \code{riskFactor}.}
}
\value{
An S3 object, instance of the class mappingTable.
}
\description{
\code{mappinTable} is the constructor for the
  S3 class mappingTable. It allows to define the market risk factors.
}
