% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{mappingTime}
\alias{mappingTime}
\title{Constructing Time Mappings}
\usage{
mappingTime(time, mapping)
}
\arguments{
\item{time}{integer value, the time to maturities.}

\item{mapping}{character value, the mapping.}
}
\value{
a data.frame with option \code{stringsAsFactors = FALSE}.
}
\description{
Constructing Time Mappings
}
\seealso{
\code{\link{marketRisk}}.
}
