% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{marketValueMargin}
\alias{marketValueMargin}
\title{Compute the Market Value Margin (MVM)}
\usage{
marketValueMargin(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters.}
}
\value{
a numeric value.
}
\description{
S3 generic method to compute the
  market value margin.
}
