% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-figures.R
\name{marketValueMargin.sstOutput}
\alias{marketValueMargin.sstOutput}
\title{Compute the Market Value Margin (MVM)}
\usage{
\method{marketValueMargin}{sstOutput}(object, nhmr = NULL, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{nhmr}{numeric value of length one. The factor for non-headgeable market risk in
market value margin computations. Default to \code{NULL}, in this case the sstOutput
  must contain this parameter. This parameter overrides \code{nhmr} in
  objects of class \code{sstOutput}.}

\item{...}{aditional parameters to be passed on to \code{expectedShortfall}.}
}
\value{
a numeric value of length one. The market value margin (MVM).
}
\description{
S3 generic method to compute the
  market value margin (MVM).
}
\seealso{
\code{\link{marketValueMargin}}.
}
