% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-general.R
\name{mvmLife}
\alias{mvmLife}
\title{MVM life computation}
\usage{
mvmLife(cashflow.table, rates, cov.mat, coc)
}
\arguments{
\item{cashflow.table}{a data.table.}

\item{rates}{a numeric vector of rates, with continuous compounding.
These should start for time to maturity 1 and go until
1 + last cashflow time to maturity.}

\item{cov.mat}{covariance matrix of life risks.}

\item{coc}{a numeric value. The cost of capital.}
}
\value{
a numeric value, the life MVM.
}
\description{
compute MVM life.
}
