% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-general.R
\name{na.rm}
\alias{na.rm}
\title{Remove Missing Values}
\usage{
na.rm(x)
}
\arguments{
\item{x}{an atomic vector.}
}
\value{
an atomic vector without NA values.
}
\description{
\code{na.rm} removes all missing values
  from a vector.
}
