% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participationRisk-base.R
\name{participationRisk}
\alias{participationRisk}
\title{Constructing a participationRisk}
\usage{
participationRisk(volatility)
}
\arguments{
\item{volatility}{positive numeric value of length one.}
}
\value{
An S3 object, instance of the class participationRisk.
}
\description{
\code{participationRisk} is the constructor for the
  S3 class participationRisk. It allows to build for participation risk parameters.
}
\examples{
# Creating a new participationRisk.
pr <- participationRisk(volatility = 0.5)

}
\seealso{
\code{\link{summary.participationRisk}}, \code{\link{print.participationRisk}},
 \code{\link{compute.participationRisk}}.
}
