% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-base.R
\name{print.asset}
\alias{print.asset}
\title{Printing an Asset with Direct Market Price}
\usage{
\method{print}{asset}(x, ...)
}
\arguments{
\item{x}{S3 object of class asset.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class asset.
}
\examples{
#' # Creating an asset.
a <- asset("equity", "USD", 1000)
# printing the asset.
print(a)

}
\seealso{
\code{\link[base]{print}}, \code{\link{asset}}.
}
