% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetForward-base.R
\name{print.assetForward}
\alias{print.assetForward}
\title{Printing an Index-Forward}
\usage{
\method{print}{assetForward}(x, ...)
}
\arguments{
\item{x}{S3 object of class assetForward.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class assetForward.
}
\examples{
# Creating an assetForward.
af <- assetForward("equity", "EUR", 1, 1000, 1200, "long")
# printing the assetForward.
print(af)

}
\seealso{
\code{\link[base]{print}}, \code{\link{assetForward}}.
}
