% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow-base.R
\name{print.cashflow}
\alias{print.cashflow}
\title{Printing a Fixed-Income-Asset}
\usage{
\method{print}{cashflow}(x, ...)
}
\arguments{
\item{x}{S3 object of class cashflow.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class cashflow.
}
\examples{
# Creating a cashflow.
ca <- cashflow(1L, "USD", "AAA", 0.5, 1000)
# printing the cashflow.
print(ca)

}
\seealso{
\code{\link[base]{print}}, \code{\link{cashflow}}
}
