% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxForward-base.R
\name{print.fxForward}
\alias{print.fxForward}
\title{Printing an FX-Forward}
\usage{
\method{print}{fxForward}(x, ...)
}
\arguments{
\item{x}{S3 object of class fxForward.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class fxForward.
}
\examples{
# Creating an fx forward.
fxf <- fxForward("USD", "EUR", 1, 1000, 1.05, "long")
# printing the fx forward.
print(fxf)

}
\seealso{
\code{\link[base]{print}}, \code{\link{fxForward}}.
}
