% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-base.R
\name{print.health}
\alias{print.health}
\title{Printing a Health Delta-Normal Term}
\usage{
\method{print}{health}(x, ...)
}
\arguments{
\item{x}{S3 object of class health.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class health.
}
\examples{
# Creating a new health item.
health1 <- health(name        = c("pandemy", "longetivity", "storno"),
                  currency    = c("EUR", "CHF", "EUR"),
                  sensitivity = c(100, 150, 130))
# printing the health item.
health1

}
\seealso{
\code{\link[base]{print}}, \code{\link{health}}.
}
