% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthRisk-base.R
\name{print.healthRisk}
\alias{print.healthRisk}
\title{Printing a HealthRisk}
\usage{
\method{print}{healthRisk}(x, ...)
}
\arguments{
\item{x}{S3 object of class healthRisk.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class healthRisk.
}
\examples{
# Creating a new healthRisk.

corr.mat <- diag(rep(1, 2))
colnames(corr.mat) <- c("invalidity", "longetivity")
rownames(corr.mat) <- colnames(corr.mat)

healthRisk1 <- healthRisk(corr.mat  = corr.mat)
# printing the healthRisk.
print(healthRisk1)

}
\seealso{
\code{\link[base]{print}}, \code{\link{healthRisk}}.
}
