% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liability-base.R
\name{print.liability}
\alias{print.liability}
\title{Printing an Insurance Liability}
\usage{
\method{print}{liability}(x, ...)
}
\arguments{
\item{x}{an S3 object of class liability.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class liability.
}
\examples{
# Creating a liability.
liab <- liability(1, "USD", 1000)
# printing the liability.
print(liab)

}
\seealso{
\code{\link[base]{print}}, \code{\link{liability}}.
}
