% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-base.R
\name{print.life}
\alias{print.life}
\title{Printing a Life Delta-Normal Remainder Term}
\usage{
\method{print}{life}(x, ...)
}
\arguments{
\item{x}{an S3 object of class life.}

\item{...}{additional parameters.}
}
\value{
None (invisible NULL).
}
\description{
print method for S3 class life.
}
\examples{
# Creating a new health item.
life1 <- life(name        = c("pandemy", "longetivity", "storno"),
              currency    = c("EUR", "CHF", "EUR"),
              sensitivity = c(100, 150, 130))
# printing the health item.
life1

}
\seealso{
\code{\link[base]{print}}, \code{\link{life}}.
}
