% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeRisk-base.R
\name{print.lifeRisk}
\alias{print.lifeRisk}
\title{Printing a LifeRisk}
\usage{
\method{print}{lifeRisk}(x, ...)
}
\arguments{
\item{x}{an S3 object of class lifeRisk.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class lifeRisk.
}
\examples{
# Creating a new lifeRisk.

corr.mat <- diag(rep(1, 2))
colnames(corr.mat) <- c("invalidity", "longetivity")
rownames(corr.mat) <- colnames(corr.mat)

lifeRisk1 <- lifeRisk(corr.mat  = corr.mat,
                      quantile = c(0.995, 0.995))
# printing the lifeRisk.
print(lifeRisk1)

}
\seealso{
\code{\link[base]{print}}, \code{\link{lifeRisk}}.
}
