% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participation-base.R
\name{print.participation}
\alias{print.participation}
\title{Printing a Participation}
\usage{
\method{print}{participation}(x, ...)
}
\arguments{
\item{x}{S3 object of class participation.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class participation.
}
\examples{
# Creating a new participation.
participation1 <- participation("USD", 1000)
# printing the participation
participation1

}
\seealso{
\code{\link[base]{print}}, \code{\link{participation}}.
}
