% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioRisk-base.R
\name{print.scenarioRisk}
\alias{print.scenarioRisk}
\title{Printing a ScenarioRisk}
\usage{
\method{print}{scenarioRisk}(x, ...)
}
\arguments{
\item{x}{S3 object of class scenarioRisk.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class scenarioRisk.
}
\examples{
# Creating a new scenarioRisk.
scenarios <- scenarioRisk(name          = c("earthquake",
                                            "real estate crash"),
                          probability   = c(0.001, 0.01),
                          currency      = c("CHF", "CHF"),
                          effect        = c(-1000, -10000))
# printing the scenarioRisk.
print(scenarios)

}
\seealso{
\code{\link{print}}, \code{\link{scenarioRisk}}.
}
