% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstModel-base.R
\name{print.sstModel}
\alias{print.sstModel}
\title{Printing a sstModel}
\usage{
\method{print}{sstModel}(x, ...)
}
\arguments{
\item{x}{S3 object of class sstModel.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class sstModel.
}
\seealso{
\code{\link[base]{print}},  \code{\link{sstModel}}.
}
