% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{print.sstOutput}
\alias{print.sstOutput}
\title{Printing a sstOutput}
\usage{
\method{print}{sstOutput}(x, ...)
}
\arguments{
\item{x}{S3 object of class sstOutput.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for S3 class sstOutput.
}
\seealso{
\code{\link[base]{print}}.
}
