% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-base.R
\name{print.standalone}
\alias{print.standalone}
\title{Printing a standalone}
\usage{
\method{print}{standalone}(x, ...)
}
\arguments{
\item{x}{S3 object of class standalone.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for the S3 class standalone.
}
\seealso{
\code{\link[base]{print}}, \code{\link{standalone}}.
}
