% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio-base.R
\name{print.summary.portfolio}
\alias{print.summary.portfolio}
\title{Printing a Summary of Portfolio}
\usage{
\method{print}{summary.portfolio}(x, ...)
}
\arguments{
\item{x}{an S3 object of class summary.portfolio.}

\item{...}{additional parameters.}
}
\value{
None (invisible NULL).
}
\description{
print method for S3 class summary.portfolio.
}
\seealso{
\code{\link[base]{print}}
}
