% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstModel-base.R
\name{print.summary.sstModel}
\alias{print.summary.sstModel}
\title{Printing a Summary of sstModel}
\usage{
\method{print}{summary.sstModel}(x, ...)
}
\arguments{
\item{x}{an S3 object of class summary.sstModel.}

\item{...}{additional parameters.}
}
\value{
None (invisible NULL).
}
\description{
print method for S3 class summary.sstModel.
}
\seealso{
\code{\link[base]{print}}
}
