% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{print.summary.sstOutput}
\alias{print.summary.sstOutput}
\title{Printing a Summary of sstOutput}
\usage{
\method{print}{summary.sstOutput}(x, ...)
}
\arguments{
\item{x}{S3 object of class summary.sstOutput.}

\item{...}{additional arguments.}
}
\value{
None (invisible NULL).
}
\description{
print method for S3 class summary.sstOutput.
}
\seealso{
\code{\link[base]{print}}.
}
