% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{rateIsIn}
\alias{rateIsIn}
\title{Rate in Object?}
\usage{
rateIsIn(object, ...)
}
\arguments{
\item{object}{an S3 object potentially containing the rate.}

\item{...}{additional parameters.}
}
\value{
a logical value.
}
\description{
S3 generic to check that the rate is in
  the object.
}
