% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-base.R
\name{rateIsIn.standalone}
\alias{rateIsIn.standalone}
\title{Rate in standalone?}
\usage{
\method{rateIsIn}{standalone}(object, currency, horizon, ...)
}
\arguments{
\item{object}{S3 object of class standalone.}

\item{currency}{character value. A currency.}

\item{horizon}{character value. An horizon.}

\item{...}{additional arguments.}
}
\value{
a logical value, is the rate in the standalone?
}
\description{
S3 generic to check that the rate is in
  the object.
}
\seealso{
\code{\link{rateIsIn}}.
}
