% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-general.R
\name{removePerfectCorr}
\alias{removePerfectCorr}
\title{Remove Perfectly Correlated Variables}
\usage{
removePerfectCorr(mat)
}
\arguments{
\item{mat}{a numeric matrix}
}
\value{
a sub matrix
}
\description{
remove perfectly correlated variables from a matrix
}
