% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-figures.R
\name{riskCapital.sstOutput}
\alias{riskCapital.sstOutput}
\title{Compute the Risk Capital (RC)}
\usage{
\method{riskCapital}{sstOutput}(object, with.scenario = F, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{with.scenario}{logical value of length one. Should the risk capital be compute
with scenario risk also?}

\item{...}{additional parameters to be passed on to
\code{expectedShortfall}.}
}
\value{
a numeric value. The risk capital (RC).
}
\description{
S3 generic method to compute the
  risk capital (RC).
}
\seealso{
\code{\link{riskCapital}}.
}
