% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-valuation.R
\name{riskFactorToExpression}
\alias{riskFactorToExpression}
\title{RiskFactor To Expression Helper}
\usage{
riskFactorToExpression(risk.factor)
}
\arguments{
\item{risk.factor}{a riskFactor object.}
}
\value{
a character value.
}
\description{
This private function creates an expression from a risk-factor.
}
