% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthRisk-simulation.R
\name{simulate.healthRisk}
\alias{simulate.healthRisk}
\title{Simulate from a HealthRisk}
\usage{
\method{simulate}{healthRisk}(object, nsim, seed = NULL, ...)
}
\arguments{
\item{object}{S3 object of class healthRisk.}

\item{nsim}{strictly positive integer value og length one. The number
of simulations.}

\item{seed}{positive integer value of length one. The seed for
reproducibility.}

\item{...}{additional asrgument to be passed to \code{rnorm}.}
}
\value{
a numeric value, the base simulations.
}
\description{
\code{simulate} is a generic S3 method for classes
  inheriting from risk. It returns a vector of risk-factor
  simulations for the corresponding risk.
}
\seealso{
\code{\link[stats]{simulate}}, \code{\link{healthRisk}}.
}
