% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeRisk-simulation.R
\name{simulate.lifeRisk}
\alias{simulate.lifeRisk}
\title{Simulate from a LifeRisk}
\usage{
\method{simulate}{lifeRisk}(object, nsim, seed = NULL, ...)
}
\arguments{
\item{object}{S3 object of class lifeRisk.}

\item{nsim}{strictly positive integer value of length one. The number
of simulations.}

\item{seed}{positive integer value of length one. The seed for
reproducibility.}

\item{...}{additional arguments to be passed to \code{rnorm}.}
}
\value{
a numeric value, the base simulations.
}
\description{
\code{simulate} is a generic S3 method for classes
  inheriting from risk. It returns a vector of risk-factor
  simulations for the corresponding risk.
}
\seealso{
\code{\link[stats]{simulate}}, \code{\link{lifeRisk}}.
}
