% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonLifeRisk-simulation.R
\name{simulate.nonLifeRisk}
\alias{simulate.nonLifeRisk}
\title{Simulate from a nonLifeRisk}
\usage{
\method{simulate}{nonLifeRisk}(object, nsim, seed = NULL, ...)
}
\arguments{
\item{object}{an S3 object of class nonLifeRisk.}

\item{nsim}{a strictly positive integer value of length one. The number
of simulations.}

\item{seed}{a positive integer value of length one. The seed for
reproducibility.}

\item{...}{additional parameters.}
}
\value{
a numeric value. The base simulations.
}
\description{
\code{simulate} is a generic S3 method for S3 classes
  inheriting from risk. It returns a vector of risk-factor
  simulations for the corresponding risk.
}
\seealso{
\code{\link[stats]{simulate}}, \code{\link{nonLifeRisk}}.
}
