% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participationRisk-simulation.R
\name{simulate.participationRisk}
\alias{simulate.participationRisk}
\title{Simulate from a participationRisk}
\usage{
\method{simulate}{participationRisk}(object, nsim, seed = NULL, ...)
}
\arguments{
\item{object}{S3 object of class participationRisk.}

\item{nsim}{strictly positive integer value og length one. The number
of simulations.}

\item{seed}{positive integer value of length one. The seed for
reproducibility.}

\item{...}{additional arguments.}
}
\value{
a numeric value. The base simulations.
}
\description{
\code{simulate} is a generic S3 method for classes
  inheriting from risk. It returns a vector of risk-factor
  simulations for the corresponding risk.
}
\seealso{
\code{\link[stats]{simulate}}, \code{\link{participationRisk}},
\code{\link{participation}}.
}
