% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioRisk-simulation.R
\name{simulate.scenarioRisk}
\alias{simulate.scenarioRisk}
\title{Simulate from a ScenarioRisk}
\usage{
\method{simulate}{scenarioRisk}(object, nsim, seed = NULL, market.risk, ...)
}
\arguments{
\item{object}{S3 object of class scenarioRisk.}

\item{nsim}{strictly positive integer value of length one. The number
of simulations.}

\item{seed}{positive integer value of length one. The seed for
reproducibility.}

\item{market.risk}{S3 object of class marketRisk created using the constructor
\code{marketRisk}.}

\item{...}{additional arguments.}
}
\value{
a numeric value, the base simulations.
}
\description{
\code{simulate} is a generic S3 method for classes
  inheriting from risk. It returns a vector of risk-factor
  simulations for the corresponding risk.
}
\seealso{
\code{\link[stats]{simulate}}, \code{\link{scenarioRisk}}.
}
