% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-general.R
\name{splitComma}
\alias{splitComma}
\title{Split Characters}
\usage{
splitComma(x, rm.spaces = T)
}
\arguments{
\item{x}{a character vector.}

\item{rm.spaces}{a logical value, should the spaces before and after commas
be deleted ?}
}
\value{
a character vector.
}
\description{
split characters by presence of `,`.
}
