% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{spreadIsIn}
\alias{spreadIsIn}
\title{Spread in Object?}
\usage{
spreadIsIn(object, ...)
}
\arguments{
\item{object}{an S3 object potentially containing the spread.}

\item{...}{additional parameters.}
}
\value{
a logical value.
}
\description{
S3 generic to check that the spread is in
  the object.
}
