% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-base.R
\name{spreadIsIn.standalone}
\alias{spreadIsIn.standalone}
\title{Spread in Standalone?}
\usage{
\method{spreadIsIn}{standalone}(object, currency, rating, ...)
}
\arguments{
\item{object}{S3 object of class standalone.}

\item{currency}{character value. A currency.}

\item{rating}{character value. The rating associated to
the spread.}

\item{...}{additional arguments.}
}
\value{
a logical value, is the spread in the standalone?
}
\description{
S3 generic to check that the spread is in
  the object.
}
\seealso{
\code{\link{spreadIsIn}}.
}
