% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstModel_check.R
\name{sstModel_check}
\alias{sstModel_check}
\title{Run checks of the packages libraries to check for potential issues.}
\usage{
sstModel_check()
}
\value{
A character vector containing all the packages' names.
}
\description{
Procedure that checks the user libraries for any package
  that can have an influence on the sstModel package running, checks if
  any of these package are built under different versions of R, and asks
  the user to update threatening packages.
}
