% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstModel_news.R
\name{sstModel_news}
\alias{sstModel_news}
\title{Display sstModel R-package News File}
\usage{
sstModel_news()
}
\description{
display the NEWS.md file to obtain information
 about new features implemented in the packages, code optimizations, changes
 of API, bug fixes, etc...
}
