% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{sstRatio}
\alias{sstRatio}
\title{Compute the Swiss Solvency Test (SST) Ratio}
\usage{
sstRatio(object, ...)
}
\arguments{
\item{object}{an S3 object.}

\item{...}{additional parameters}
}
\value{
a numeric value.
}
\description{
S3 generic method to compute the
  sst ratio.
}
