% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-figures.R
\name{sstRatio.sstOutput}
\alias{sstRatio.sstOutput}
\title{Compute the Swiss Solvency Test (SST) Ratio}
\usage{
\method{sstRatio}{sstOutput}(object, with.scenario = F, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{with.scenario}{logical value of length one. Should the target capital be compute
with scenario risk also?}

\item{...}{additional parameters to be passed on to
\code{marketValueMargin} and \code{riskCapital}.}
}
\value{
a numeric value. The Swiss Solvency Test Ratio.
}
\description{
S3 generic method to compute the
  sst ratio.
}
\seealso{
\code{\link{sstRatio}}.
}
