% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{standaloneExpectedShortfall}
\alias{standaloneExpectedShortfall}
\title{Compute expected shortfall for standalone risk by reference}
\usage{
standaloneExpectedShortfall(object, ...)
}
\arguments{
\item{object}{an S3 object of class sstOutput.}

\item{...}{additional parameters passed to \code{expectedShortfall}.}
}
\value{
a numeric value, the expected shortfall.
}
\description{
S3 generic method to compute expected shortfall of
a standalone risk.
}
\seealso{
\code{\link{getDrbc}}
}
