% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-base.R
\name{summary.asset}
\alias{summary.asset}
\title{Summarizing an Asset with Direct Market Price}
\usage{
\method{summary}{asset}(object, ...)
}
\arguments{
\item{object}{S3 object of class asset.}

\item{...}{additional arguments.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class asset.
}
\examples{
# Creating an asset.
a <- asset("equity", "USD", 1000)
# summarizing the asset.
summary(a)

}
\seealso{
\code{\link[base]{summary}}, \code{\link{asset}}.
}
