% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow-base.R
\name{summary.cashflow}
\alias{summary.cashflow}
\title{Summarizing a Fixed-Income-Asset}
\usage{
\method{summary}{cashflow}(object, ...)
}
\arguments{
\item{object}{S3 object of class cashflow.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class cashflow.
}
\examples{
# Creating a cashflow.
ca <- cashflow(1L, "USD", "AAA", 0.1, 1000)
# summarizing the cashflow.
summary(ca)

}
\seealso{
\code{\link[base]{summary}}, \code{\link{cashflow}}
}
