% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-base.R
\name{summary.delta}
\alias{summary.delta}
\title{Summarizing a Delta-Normal Remainder Term}
\usage{
\method{summary}{delta}(object, ...)
}
\arguments{
\item{object}{S3 object of class delta.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
an S3 object,instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for S3 class delta.
}
\examples{
# Creating a new delta.
delta1 <- delta(name        = c("equity", "2YCHF", "EURCHF"),
                currency    = c("EUR", "CHF", "EUR"),
                sensitivity = c(100, 150, 130))
# summarizing the delta.
summary(delta1)

}
\seealso{
\code{\link[base]{summary}}, \code{\link{delta}}.
}
