% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-base.R
\name{summary.health}
\alias{summary.health}
\title{Summarizing a Health Delta-Normal Term}
\usage{
\method{summary}{health}(object, ...)
}
\arguments{
\item{object}{3 S3 object of class health.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class health.
}
\examples{
# Creating a new health item.
health1 <- health(name        = c("pandemy", "longetivity", "storno"),
                  currency    = c("EUR", "CHF", "EUR"),
                  sensitivity = c(100, 150, 130))
# summarizing the health item.
summary(health1)

}
\seealso{
\code{\link[base]{summary}}, \code{\link{health}}.
}
