% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthRisk-base.R
\name{summary.healthRisk}
\alias{summary.healthRisk}
\title{Summarizing a HealthRisk}
\usage{
\method{summary}{healthRisk}(object, ...)
}
\arguments{
\item{object}{S3 object of class healthRisk.}

\item{...}{additional arguments.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class healthRisk.
}
\examples{
# Creating a new healthRisk.

corr.mat <- diag(rep(1, 2))
colnames(corr.mat) <- c("invalidity", "longetivity")
rownames(corr.mat) <- colnames(corr.mat)

healthRisk1 <- healthRisk(corr.mat  = corr.mat)
# summarizing the healthRisk.
summary(healthRisk1)

}
\seealso{
\code{\link[base]{summary}}, \code{\link{healthRisk}}.
}
