% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liability-base.R
\name{summary.liability}
\alias{summary.liability}
\title{Summarizing an Insurance Liability}
\usage{
\method{summary}{liability}(object, ...)
}
\arguments{
\item{object}{S3 object of class liability.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class liability.
}
\examples{
# Creating a liability.
liab <- liability(1, "USD", 1000)
# summarizing the liability.
summary(liab)

}
\seealso{
\code{\link[base]{summary}}, \code{\link{liability}}.
}
