% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-base.R
\name{summary.life}
\alias{summary.life}
\title{Summarizing a Life Delta-Normal Remainder Term}
\usage{
\method{summary}{life}(object, ...)
}
\arguments{
\item{object}{S3 object of class life.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class life.
}
\examples{
# Creating a new life item.
life1 <-   life(name        = c("pandemy", "longetivity", "storno"),
                currency    = c("EUR", "CHF", "EUR"),
                sensitivity = c(100, 150, 130))
# summarizing the life item.
summary(life1)

}
\seealso{
\code{\link[base]{summary}}, \code{\link{life}}
}
