% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketRisk-base.R
\name{summary.marketRisk}
\alias{summary.marketRisk}
\title{Summarizing a marketRisk}
\usage{
\method{summary}{marketRisk}(object, ...)
}
\arguments{
\item{object}{S3 object of class marketRisk.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
a table with names:
  \itemize{
    \item \code{base risk-factors}: the number of base risk-factors
      in the marketRisk.
    \item \code{scaled risk-factors}: the number of scaled risk-factors
      in the marketRisk.
    \item \code{base currency}: the base currency.
  }
}
\description{
summary method for S3 class marketRisk.
}
\seealso{
\code{\link[base]{summary}}, \code{\link{marketRisk}}.
}
