% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participation-base.R
\name{summary.participation}
\alias{summary.participation}
\title{Summarizing a Participation}
\usage{
\method{summary}{participation}(object, ...)
}
\arguments{
\item{object}{S3 object of class participation.}

\item{...}{additional arguments.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class participation.
}
\examples{
# Creating a new participation.
participation1 <- participation("USD", 1000)
# summarizing the participation
summary(participation1)

}
\seealso{
\code{\link{summary}}, \code{\link{participation}}.
}
