% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participationRisk-base.R
\name{summary.participationRisk}
\alias{summary.participationRisk}
\title{Summarizing a participationRisk}
\usage{
\method{summary}{participationRisk}(object, ...)
}
\arguments{
\item{object}{S3 object of class participationRisk.}

\item{...}{additional arguments.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class participationRisk.
}
\examples{
# Creating a new participationRisk.
pr <- participationRisk(volatility = 0.5)
# summarizing the participationRisk.
summary(pr)

}
\seealso{
\code{\link[base]{summary}}, \code{\link{participationRisk}}.
}
