% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioRisk-base.R
\name{summary.scenarioRisk}
\alias{summary.scenarioRisk}
\title{Summarizing a ScenarioRisk}
\usage{
\method{summary}{scenarioRisk}(object, ...)
}
\arguments{
\item{object}{S3 object of class scenarioRisk.}

\item{...}{additional arguments.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for the S3 class scenarioRisk.
}
\examples{
# Creating a new scenarioRisk.
scenarios <- scenarioRisk(name          = c("earthquake",
                                            "real estate crash"),
                          probability   = c(0.001, 0.01),
                          currency      = c("CHF", "CHF"),
                          effect        = c(1000, 10000))
# summarizing the scenarioRisk.
summary(scenarios)

}
\seealso{
\code{\link{summary}}, \code{\link{scenarioRisk}}.
}
