% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstModel-base.R
\name{summary.sstModel}
\alias{summary.sstModel}
\title{Summarizing an sstModel}
\usage{
\method{summary}{sstModel}(object, ...)
}
\arguments{
\item{object}{S3 object of class sstModel.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
an S3 object, instance of class \code{summary.sstModel}.
}
\description{
summary method for the S3 sstModel.
}
\seealso{
\code{\link[base]{summary}}, \code{\link{sstModel}}.
}
