% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{summary.sstOutput}
\alias{summary.sstOutput}
\title{Summarizing a sstOutput}
\usage{
\method{summary}{sstOutput}(object, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{...}{additional arguments to be passed to
\code{marketValueMargin}, \code{riskCapital},
\code{targetCapital}, \code{sstRatio},
\code{expectedShortfall}. It allows to modify parameters
\code{nhmr} for market value margin computations, \code{alpha}
and \code{sup} for all expected shortfall computations
with \code{expectedShortfall}.}
}
\value{
an S3 object, instance of class \code{c("summaryDefault", "table")}.
}
\description{
summary method for S3 class sstOutput.
}
\seealso{
\code{\link[base]{summary}}.
}
