% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketItem-parser.R
\name{tableToAssetForward}
\alias{tableToAssetForward}
\title{Parsing a table to a list of assetForward}
\usage{
tableToAssetForward(table)
}
\arguments{
\item{table}{a data.frame.}
}
\value{
a list of object of class assetForward.
}
\description{
internal helper for parsing.
}
\seealso{
\code{\link{assetForward}}.
}
