% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketItem-parser.R
\name{tableToAssets}
\alias{tableToAssets}
\title{Parsing a table to a list of asset}
\usage{
tableToAssets(table)
}
\arguments{
\item{table}{a data.frame.}
}
\value{
a list of object of class asset.
}
\description{
internal helper for parsing.
}
\seealso{
\code{\link{asset}}.
}
