% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketItem-parser.R
\name{tableToCashflow}
\alias{tableToCashflow}
\title{Parsing a table to a list of cashflow}
\usage{
tableToCashflow(table)
}
\arguments{
\item{table}{a data.frame.}
}
\value{
a list of object of class liability.
}
\description{
internal helper for parsing.
}
\seealso{
\code{\link{cashflow}}.
}
