% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketItem-parser.R
\name{tableToFxForward}
\alias{tableToFxForward}
\title{Parsing a table to a list of fxForward}
\usage{
tableToFxForward(table)
}
\arguments{
\item{table}{a data.frame.}
}
\value{
a list of object of class assetForward.
}
\description{
internal helper for parsing.
}
\seealso{
\code{\link{fxForward}}.
}
