% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketItem-parser.R
\name{tableToLiability}
\alias{tableToLiability}
\title{Parsing a table to a list of liability}
\usage{
tableToLiability(table)
}
\arguments{
\item{table}{a data.frame.}
}
\value{
a list of object of class liability.
}
\description{
internal helper for parsing.
}
\seealso{
\code{\link{liability}}.
}
