% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{targetCapital}
\alias{targetCapital}
\title{Target Capital}
\usage{
targetCapital(object, ...)
}
\arguments{
\item{object}{an S3 object from which to obtain the target capital.}

\item{...}{additional parameters.}
}
\value{
information about target capital.
}
\description{
\code{targetCapital} is a generic S3 method for S3 classes
  from which target capital can be provided.
}
