% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-figures.R
\name{targetCapital.sstOutput}
\alias{targetCapital.sstOutput}
\title{Compute the Target Capital (TC)}
\usage{
\method{targetCapital}{sstOutput}(object, with.scenario = F, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{with.scenario}{logical value of length one. Should the target capital be compute
with scenario risk also?}

\item{...}{additional parameters to be passed on to
\code{marketValueMargin} and \code{riskCapital}.}
}
\value{
a numeric value. The target Capital (TC).
}
\description{
\code{targetCapital} is a generic S3 method for S3 classes
  from which target capital can be provided.
}
\seealso{
\code{\link{targetCapital}}.
}
