% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{translate}
\alias{translate}
\title{translate}
\usage{
translate(object, ...)
}
\arguments{
\item{object}{an S3 object to translate the fields.}

\item{...}{additional parameters.}
}
\value{
a character vector.
}
\description{
\code{translate} is a generic S3 method for translating
  variable names to understandable sentences.
}
