% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{translate.sstOutput}
\alias{translate.sstOutput}
\title{Translation of Fields of sstOutput}
\usage{
\method{translate}{sstOutput}(object, ...)
}
\arguments{
\item{object}{S3 object of class sstOutput.}

\item{...}{additional arguments.}
}
\value{
a named character vector. The values correspond
  to the columns of \code{object} and the names to their
  translation to humanly readable titles.
}
\description{
translate S3 method for sstOutput. This method allow
 to translate code-related naming convention to human-understandable
 names.
}
\seealso{
\code{\link[base]{summary}}.
}
