% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-valuation.R
\name{valExpression.asset}
\alias{valExpression.asset}
\title{Building the Valuation Expression for Asset with Direct Market
Price}
\usage{
\method{valExpression}{asset}(object, market.risk, standalone = NULL, ...)
}
\arguments{
\item{object}{S3 object of class asset.}

\item{market.risk}{S3 object of class marketRisk.}

\item{standalone}{S3 object of class standalone.}

\item{...}{additional arguments.}
}
\value{
character value of length one, the expression representing the valuation
  of the asset position.
}
\description{
\code{valExpression} is a generic S3 method for classes
  inheriting from item. It returns the valuation expression.
}
\seealso{
\code{\link{valExpression}}, \code{\link{asset}},
  \code{\link{marketRisk}}, \code{\link{standalone}}.
}
