% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetForward-valuation.R
\name{valExpression.assetForward}
\alias{valExpression.assetForward}
\title{Building the Valuation Expression for an Index-Forward}
\usage{
\method{valExpression}{assetForward}(object, market.risk, standalone = NULL,
  ...)
}
\arguments{
\item{object}{S3 object of class assetForward.}

\item{market.risk}{S3 object of class marketRisk created using
\code{marketRisk}.}

\item{standalone}{S3 object of class standalone.}

\item{...}{additional arguments.}
}
\value{
a character value. The expression representing the valuation
  of the index-forward position.
}
\description{
\code{valExpression} is a generic S3 method for classes
  inheriting from item. It returns the valuation expression.
}
\seealso{
\code{\link{valExpression}}, \code{\link{assetForward}}.
}
