% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-valuation.R
\name{valExpression.delta}
\alias{valExpression.delta}
\title{Building the Valuation Expression for a Market
Delta-Normal Remainder Term}
\usage{
\method{valExpression}{delta}(object, market.risk, standalone = NULL, ...)
}
\arguments{
\item{object}{S3 object of class delta.}

\item{market.risk}{S3 object of class marketRisk created using
\code{marketRisk}.}

\item{standalone}{S3 object of class standalone.}

\item{...}{additional arguments.}
}
\value{
a character value. The expression representing the valuation
  of the delta remainder term.
}
\description{
\code{valExpression} is a generic S3 method for classes
  inheriting from item. It returns the valuation expression.
}
\seealso{
\code{\link{valExpression}}, \code{\link{delta}}.
}
