% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxForward-valuation.R
\name{valExpression.fxForward}
\alias{valExpression.fxForward}
\title{Building the Valuation Expression for a FX-Forward Position}
\usage{
\method{valExpression}{fxForward}(object, market.risk, standalone = NULL, ...)
}
\arguments{
\item{object}{S3 object of class fxForward.}

\item{market.risk}{S3 object of class marketRisk created using
\code{marketRisk}.}

\item{standalone}{S3 object of class standalone.}

\item{...}{additional arguments.}
}
\value{
a character value. The expression representing the valuation
  of the fx forward position.
}
\description{
\code{valExpression} is a generic S3 method for classes
  inheriting from item. It returns the valuation expression.
}
\seealso{
\code{\link{valExpression}}, \code{\link{fxForward}}.
}
