% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-valuation.R
\name{valExpression.health}
\alias{valExpression.health}
\title{Building the Valuation Expression for a Health Item}
\usage{
\method{valExpression}{health}(object, market.risk, health.risk, ...)
}
\arguments{
\item{object}{S3 object of class health.}

\item{market.risk}{S3 object of class marketRisk created using the constructor
\code{marketRisk}.}

\item{health.risk}{S3 object of class healthRisk created using the constructor
\code{healthRisk}.}

\item{...}{additional arguments.}
}
\value{
a character value. The expression representing the valuation
  of the health item.
}
\description{
\code{valExpression} is a generic S3 method for classes
  inheriting from item. It returns the valuation expression.
}
\seealso{
\code{\link{valExpression}}, \code{\link{health}}.
}
