% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liability-valuation.R
\name{valExpression.liability}
\alias{valExpression.liability}
\title{Building the Valuation Expression for an Insurance Liability}
\usage{
\method{valExpression}{liability}(object, market.risk, standalone = NULL, ...)
}
\arguments{
\item{object}{S3 object of class liability.}

\item{market.risk}{S3 object of class marketRisk created using the constructor
\code{marketRisk}.}

\item{standalone}{S3 object of class standalone.}

\item{...}{additional arguments.}
}
\value{
a character value. The expression representing the valuation
  of the liability position.
}
\description{
\code{valExpression} is a generic S3 method for S3 classes
  inheriting from item. It returns the valuation expression.
}
\seealso{
\code{\link{valExpression}}, \code{\link{liability}},
  \code{\link{marketRisk}}, \code{\link{standalone}}.
}
