% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-valuation.R
\name{valExpression.life}
\alias{valExpression.life}
\title{Building the Valuation Expression for a Life Item}
\usage{
\method{valExpression}{life}(object, market.risk, life.risk, ...)
}
\arguments{
\item{object}{S3 object of class life.}

\item{market.risk}{S3 object of class marketRisk created using
\code{marketRisk}.}

\item{life.risk}{S3 bject of class lifeRisk created using
\code{lifeRisk}.}

\item{...}{additional arguments.}
}
\value{
a character value. The expression representing the valuation
  of the life item.
}
\description{
\code{valExpression} is a generic S3 method for classes
  inheriting from item. It returns the valuation expression.
}
\seealso{
\code{\link{valExpression}}, \code{\link{life}}.
}
