% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{valExpression}
\alias{valExpression}
\title{Valuation Expression}
\usage{
valExpression(object, ...)
}
\arguments{
\item{object}{an S3 object from which to construct a
valuation expression.}

\item{...}{additional parameters.}
}
\value{
a character value.
}
\description{
\code{valExpression} is a generic S3 method for S3 classes
  inheriting from item. It returns the valuation expression.
}
