% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-valuation.R
\name{valFunction.delta}
\alias{valFunction.delta}
\title{Building the Valuation Function for a Market
Delta-Normal Remainder Term}
\usage{
\method{valFunction}{delta}(object, market.risk, ...)
}
\arguments{
\item{object}{S3 object of class delta.}

\item{market.risk}{S3 object of class marketRisk created using
\code{marketRisk}.}

\item{...}{additional arguments.}
}
\value{
a function with arguments:
          \itemize{
            \item \code{x}: a matrix of simulations (numeric values) with named columns corresponding
              exactly to the name of base risk-factors in \code{marketRisk} keeping the
              same order, or an unnamed vector of simulations (numeric values) keeping the same
              ordering of base risk-factors as in \code{marketRisk}.
          }
}
\description{
\code{valFunction} is a generic S3 method for classes
  inheriting from item. It returns the valuation function.
}
\seealso{
\code{\link{valFunction}}, \code{\link{delta}}.
}
