% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{valFunction}
\alias{valFunction}
\title{Valuation Function}
\usage{
valFunction(object, ...)
}
\arguments{
\item{object}{an S3 object from which to construct a
valuation function.}

\item{...}{additional parameters.}
}
\value{
a function.
}
\description{
\code{valFunction} is a generic S3 method for S3 classes
  inheriting from item. It returns the valuation function.
}
