% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxForward-valuation.R
\name{valInfo.fxForward}
\alias{valInfo.fxForward}
\title{Providing Information for FX-Forward Valuation from a marketRisk}
\usage{
\method{valInfo}{fxForward}(object, market.risk, standalone = NULL, ...)
}
\arguments{
\item{object}{S3 object of class fxForward.}

\item{market.risk}{S3 object of class marketRisk created using
\code{marketRisk}.}

\item{standalone}{S3 object of class standalone.}

\item{...}{additional arguments.}
}
\value{
A list with the following elements:
\itemize{
  \item \code{floating.term}: a liability item. The liability
    term containing the fx rate risk.
  \item \code{fixed.term}: a liability item. The liability term
    containing the fixed exchange rate.
}
}
\description{
\code{valInfo} is a generic S3 method for classes
  inheriting from item. It returns sufficient information for the
  creation of the valuation function of the item.
}
