% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericDispatch.R
\name{valInfo}
\alias{valInfo}
\title{Providing Valuation Information}
\usage{
valInfo(object, ...)
}
\arguments{
\item{object}{an S3 object from which to extract information.}

\item{...}{additional parameters.}
}
\value{
a list.
}
\description{
\code{valInfo} is a generic S3 method for S3 classes
  inheriting from item. It returns sufficient information for the
  creation of the valuation function of the item.
}
