% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-sstModel-computation.R
\name{valueAtRisk}
\alias{valueAtRisk}
\title{Compute the Value-at-Risk}
\usage{
valueAtRisk(x, alpha = 0.005)
}
\arguments{
\item{x}{a numeric vector. The vector from which to compute the value-at-risk.}

\item{alpha}{numeric value, the alpha-Value-at-Risk, must take values
between 0 and 1. Please note that we consider value-at-risk here to be equivalent
to the alpha-quantiles of \code{x}.}
}
\value{
a numeric value. The value-at-risk.
}
\description{
function to compute the alpha-Value-at-Risk of a vector.
}
