% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-sstModel-computation.R
\name{volaToExpectedShortfall}
\alias{volaToExpectedShortfall}
\title{Transform normal volatility in expected shortfall}
\usage{
volaToExpectedShortfall(x, alpha = 0.01, sup = F, ...)
}
\arguments{
\item{x}{a numeric vector of positive volatilities.}

\item{alpha}{a numeric value. The alpha-Expected Shortfall, must take values
between 0 and 1.}

\item{sup}{a logical value. If \code{TRUE} the function returns the upper
expected shortfall and otherwise the lower. Default is set to \code{FALSE}.}

\item{...}{additional parameters.}
}
\value{
a numeric vector, the expected shortfalls.
}
\description{
function to compute expected shortfall from volatility for
  normal random variables.
}
