% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstOutput-base.R
\name{write.sstOutput}
\alias{write.sstOutput}
\title{Writing a sstOutput into a fundamental data sheet}
\usage{
write.sstOutput(object, path, keep = NULL, new.names = NULL, ...)
}
\arguments{
\item{object}{S3 object of class sstOuput.}

\item{path}{the complete path to the created .xlsx file.}

\item{keep}{character value, by default set to \code{NULL}.
The names of the columns of the field \code{$simulations} of the
sstOutput to save additionally to the fundamental data sheet.}

\item{new.names}{character value, replacement names for the
columns to keep.}

\item{...}{additional arguments to be passed on to
\code{summary.sstOutput}.}
}
\value{
None (only used for side-effects).
}
\description{
write an sstOutput in a .xlsx file.
}
\note{
This function is an interface that writes the output of
  \code{summary.sstOutput} into an excel file.
}
\seealso{
\code{\link{summary}}.
}
